requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
werkzeug==2.3.7
bleach==6.0.0
markdown==3.5.1
gunicorn==21.2.0